+++ src/data/protocols/db/cat/add-cat.repository.ts
@@ -1,0 +1,6 @@
+import { Cat } from '@/infra/db/entities/cat/cat.entity';
+import { CreateCatDto } from '@/modules/cats/dtos/create-cat/create-cat.dto';
+
+export interface AddCatRepository {
+  add: (createCatDto: CreateCatDto) => Promise<Cat>;
+}