+++ src/data/protocols/db/cat/find-by-id.repository.ts
@@ -1,0 +1,5 @@
+import { Cat } from '@/infra/db/entities/cat/cat.entity';
+
+export interface FindByIdRepository {
+  findById: (id: string) => Promise<Cat>;
+}