+++ src/docs/swagger-config.ts
@@ -5,11 +5,12 @@ export const swaggerConfig = (app: NestExpressApplication): void => {
   const config = new DocumentBuilder()
     .setTitle('Pandora Example')
     .setDescription('The Pandora API description')
-    .setVersion('0.0.1')
+    .setVersion('1.0.1')
     .addTag('users')
     .addTag('auth')
     .addTag('health')
     .addTag('roles')
+    .addTag('cats')
     .addBearerAuth()
     .build();
   const document = SwaggerModule.createDocument(app, config);