+++ src/modules/cats/cats.module.ts
@@ -1,0 +1,25 @@
+import { Cat } from '@/infra/db/entities/cat/cat.entity';
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { CatsRepository } from '@/modules/cats/repositories/cats.repository';
+import { AddCatService } from './services/add-cat/add-cat.service';
+import { CatsController } from './controllers/cats.controller';
+import { LoadUserByRoleService } from '../roles/services/load-user-by-role/load-user-by-role.service';
+import { LoadUserByIdService } from '../users/services/load-user-by-id/load-user-by-id.service';
+import { RoleRepository } from '../roles/repositories/role.repository';
+import { UserRepository } from '../users/repositories/user.repository';
+import { FindCountService } from '@/modules/cats/services/find-count/find-count.service';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([
+      Cat,
+      CatsRepository,
+      RoleRepository,
+      UserRepository,
+    ]),
+  ],
+  providers: [AddCatService, LoadUserByRoleService, LoadUserByIdService, FindCountService],
+  controllers: [CatsController],
+})
+export class CatsModule {}