+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Add new version 1.0.1 in Swagger
+- Add Sentry and SentryInterceptor
+- Add CatsModule
+- Add VerifyToken
+- Add findById, findByName and findCatAndCount in catsRepository
+
+### Removed
+- Remove node-fetch and module.hot in Main
 
 ## Release [v1.0.1] - 2021-07-17